# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Subway Guatemala Customer API",
        "description": "API REST para la aplicación móvil de clientes de Subway Guatemala. Incluye autenticación multi-canal (email/password, Google OAuth, Apple Sign-In), gestión de perfil, dispositivos FCM y futuras funcionalidades de pedidos y lealtad.",
        "contact": {
            "email": "dev@subwayguatemala.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost",
            "description": "Application Server"
        }
    ],
    "paths": {
...
        "/api/v1/devices": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "List customer devices",
                "description": "Returns all active devices for authenticated customer. Current device is flagged with is_current_device=true.",
                "operationId": "afcbc146bf93b130969374c0c29f42f6",
                "responses": {
                    "200": {
                        "description": "Devices retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CustomerDevice"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/devices/register": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Register or update device",
                "description": "Registers new device or updates existing device with FCM token. Associates device with current Sanctum token for session management.",
                "operationId": "0743195c75a8fccfdf9875291bd6d8c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fcm_token",
                                    "device_type"
                                ],
                                "properties": {
                                    "fcm_token": {
                                        "description": "Firebase Cloud Messaging token",
                                        "type": "string",
                                        "example": "fKw8h4Xj..."
                                    },
                                    "device_identifier": {
                                        "description": "Unique device identifier (auto-generated if not provided)",
                                        "type": "string",
                                        "example": "ABC123DEF456",
                                        "nullable": true
                                    },
                                    "device_type": {
                                        "description": "Device platform",
                                        "type": "string",
                                        "enum": [
                                            "ios",
                                            "android",
                                            "web"
                                        ],
                                        "example": "ios"
                                    },
                                    "device_name": {
                                        "description": "Human-readable device name",
                                        "type": "string",
                                        "example": "iPhone 14 Pro",
                                        "nullable": true
                                    },
                                    "device_model": {
                                        "description": "Device model identifier",
                                        "type": "string",
                                        "example": "iPhone15,2",
                                        "nullable": true
                                    },
                                    "app_version": {
                                        "description": "App version number",
                                        "type": "string",
                                        "example": "1.0.5",
                                        "nullable": true
                                    },
                                    "os_version": {
                                        "description": "OS version number",
                                        "type": "string",
                                        "example": "17.2",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Device registered or updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dispositivo registrado exitosamente."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerDevice"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/devices/{device}": {
            "delete": {
                "tags": [
                    "Devices"
                ],
                "summary": "Deactivate device",
                "description": "Marks device as inactive (soft deactivate). Device will no longer receive push notifications.",
                "operationId": "a9545ae592649836835794a3421d3551",
                "parameters": [
                    {
                        "name": "device",
                        "in": "path",
                        "description": "Device ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device deactivated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dispositivo desactivado exitosamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - device belongs to another customer"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomerDevice": {
                "title": "CustomerDevice",
                "description": "Customer device registered for push notifications",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sanctum_token_id": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "fcm_token": {
                        "type": "string",
                        "example": "fKw8h4Xj..."
                    },
                    "device_identifier": {
                        "type": "string",
                        "example": "ABC123DEF456"
                    },
                    "device_type": {
                        "type": "string",
                        "enum": [
                            "ios",
                            "android",
                            "web"
                        ],
                        "example": "ios"
                    },
                    "device_name": {
                        "type": "string",
                        "example": "iPhone 14 Pro",
                        "nullable": true
                    },
                    "device_model": {
                        "type": "string",
                        "example": "iPhone15,2",
                        "nullable": true
                    },
                    "app_version": {
                        "type": "string",
                        "example": "1.0.5",
                        "nullable": true
                    },
                    "os_version": {
                        "type": "string",
                        "example": "17.2",
                        "nullable": true
                    },
                    "last_used_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-20T15:45:00Z",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_current_device": {
                        "description": "True if this device is associated with current token",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-10T08:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-20T15:45:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token. Obtain from /api/v1/auth/login, /api/v1/auth/register, or OAuth endpoints.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}